+++ packages/react/src/components/video/EncryptedVideoPlayer.tsx
@@ -1,8 +1,9 @@
-import React, { VideoHTMLAttributes, useMemo } from 'react';
+import { NFT } from '@cere/media-sdk-client';
+import React, { VideoHTMLAttributes, useState, useEffect } from 'react';
 
-import { useMediaClient, useServerSideUrl } from '../../hooks';
+import { useEncryptedContent, useMediaClient, useNftMetadata, useServerSideUrl } from '../../hooks';
 
-import { HlsEncryptionLoader } from './HlsEncryptionLoader';
+import { createHlsEncryptionLoader } from './HlsEncryptionLoader';
 import { IosVideoPlayer } from './IosVideoPlayer';
 import { VideoPlayer } from './VideoPlayer';
 
@@ -10,7 +11,7 @@ export interface EncryptedVideoPlayerProps {
   src: string;
   collectionAddress: string;
   nftId: number;
-  assetIndex: number;
+  assetIndex?: number;
   serverSide?: boolean;
   className?: string;
   loadingComponent?: React.ReactNode;
@@ -21,23 +22,37 @@ export const EncryptedVideoPlayer = ({
   src,
   collectionAddress,
   nftId,
-  assetIndex,
+  assetIndex = 0,
   serverSide,
   ...props
 }: EncryptedVideoPlayerProps) => {
   const { client } = useMediaClient();
   const { url } = useServerSideUrl({ src, collectionAddress, nftId });
+  const { metadata } = useNftMetadata(collectionAddress, nftId);
+  const nft = {
+    collection: { address: collectionAddress },
+    nftId,
+  } as NFT;
+  const { asset } = useEncryptedContent(nft, metadata!, assetIndex);
+  const assetCid = asset.asset.split('/').pop();
 
-  const loader = useMemo(() => {
-    return client
-      ? HlsEncryptionLoader.create({
+  const [loader, setLoader] = useState<any>(null);
+
+  useEffect(() => {
+    const initializeLoader = async () => {
+      if (client) {
+        const loaderInstance = await createHlsEncryptionLoader({
           collectionAddress,
           nftId,
-          assetId: `asset-${assetIndex}`,
+          assetId: `asset-${assetCid}`,
           client,
-        })
-      : undefined;
-  }, [collectionAddress, nftId, assetIndex, client]);
+        });
+        setLoader(() => loaderInstance); // Set the loader instance once it's ready
+      }
+    };
+
+    initializeLoader();
+  }, [client]);
 
   if (!client) {
     return <></>;