+++ packages/react/src/components/video/VideoPlayer.tsx
@@ -2,14 +2,13 @@ import './plyr.css';
 import './styles.css';
 
 import clsx from 'clsx';
-import Hls, { HlsConfig } from 'hls.js';
-import Plyr from 'plyr';
+import type { Level } from 'hls.js';
 import { VideoHTMLAttributes, useEffect, useMemo, useRef, useState } from 'react';
 
 interface VideoPlayerProps {
   src: string;
   hlsEnabled?: boolean;
-  loader?: HlsConfig['loader'];
+  loader?: any; // Adjusted to accommodate dynamic import
   className?: string;
   loadingComponent?: React.ReactNode;
   type?: string;
@@ -19,7 +18,7 @@ interface VideoPlayerProps {
 export const VideoPlayer = ({
   src,
   hlsEnabled = true,
-  loader = Hls.DefaultConfig.loader,
+  loader,
   className,
   loadingComponent,
   type,
@@ -28,37 +27,49 @@ export const VideoPlayer = ({
   const wrapperRef = useRef<HTMLDivElement>(null);
   const playerRef = useRef<HTMLVideoElement | null>(null);
   const [isLoading, setIsLoading] = useState<boolean>(true);
+  const [hlsInstance, setHlsInstance] = useState<any>(null);
+  const [Plyr, setPlyr] = useState<any>(null);
 
-  const isVideoSupported = useMemo(() => (hlsEnabled ? Hls.isSupported() : true), [hlsEnabled]);
+  const isVideoSupported = useMemo(() => (hlsEnabled ? hlsInstance?.isSupported() : true), [hlsEnabled, hlsInstance]);
 
   useEffect(() => {
-    if (playerRef.current) return;
+    const loadDependencies = async () => {
+      if (hlsEnabled) {
+        const { default: Hls } = await import('hls.js');
+        setHlsInstance(() => Hls); // Ensure Hls is correctly set as a constructor function
+      }
+      const { default: PlyrModule } = await import('plyr');
+      setPlyr(() => PlyrModule);
+    };
+
+    loadDependencies();
+  }, [hlsEnabled]);
+
+  useEffect(() => {
+    if (!isVideoSupported || !Plyr) return;
 
     const videoWrapper = wrapperRef.current;
     if (!videoWrapper) return;
 
-    const hlsOptions: Partial<HlsConfig> = {
-      fragLoadingTimeOut: 30_000,
-      manifestLoadingTimeOut: 10_000,
-      levelLoadingTimeOut: 15_000,
-      loader,
-    };
-
-    const hls = new Hls(hlsOptions);
     const plyrOptions: Plyr.Options = {
       autoplay: videoOverrides.autoPlay,
     };
 
-    if (!isVideoSupported) {
-      return;
-    }
+    const initializeHls = () => {
+      const hlsOptions: Partial<any> = {
+        fragLoadingTimeOut: 30_000,
+        manifestLoadingTimeOut: 10_000,
+        levelLoadingTimeOut: 15_000,
+        loader: loader || hlsInstance.DefaultConfig.loader,
+      };
 
-    if (hlsEnabled) {
+      const hls = new hlsInstance(hlsOptions);
+
       const updateQuality = (newQuality: number) => {
         if (newQuality === 0) {
           hls.currentLevel = -1;
         } else {
-          hls.levels.forEach((level, levelIndex) => {
+          hls.levels.forEach((level: Level, levelIndex: number) => {
             if (level.height === newQuality) {
               hls.currentLevel = levelIndex;
             }
@@ -66,8 +77,8 @@ export const VideoPlayer = ({
         }
       };
 
-      hls.on(Hls.Events.MANIFEST_PARSED, () => {
-        const availableQualities = hls.levels.map((l) => l.height);
+      hls.on(hlsInstance.Events.MANIFEST_PARSED, () => {
+        const availableQualities = hls.levels.map((l: Level) => l.height);
         plyrOptions.quality = {
           default: -1, // auto
           options: availableQualities,
@@ -84,9 +95,11 @@ export const VideoPlayer = ({
         player.on('canplaythrough', () => setIsLoading(false));
       });
 
-      hls.on(Hls.Events.FRAG_LOADED, () => setIsLoading(false));
+      hls.on(hlsInstance.Events.FRAG_LOADED, () => setIsLoading(false));
       hls.loadSource(src);
-    } else {
+    };
+
+    const initializeVideo = () => {
       const video = document.createElement('video');
       const source = document.createElement('source');
 
@@ -107,16 +120,24 @@ export const VideoPlayer = ({
       video.addEventListener('error', () => setIsLoading(false));
       video.addEventListener('stalled', () => setIsLoading(false));
       video.addEventListener('suspend', () => setIsLoading(false));
+    };
+
+    if (hlsEnabled && hlsInstance) {
+      initializeHls();
+    } else {
+      initializeVideo();
     }
 
     return () => {
-      hls.destroy();
+      if (hlsEnabled && hlsInstance && typeof hlsInstance.destroy === 'function') {
+        hlsInstance.destroy();
+      }
       if (playerRef.current) {
         playerRef.current.remove();
       }
       playerRef.current = null;
     };
-  }, [isVideoSupported, loader, src, wrapperRef]);
+  }, [hlsInstance, hlsEnabled, isVideoSupported, loader, src, wrapperRef, videoOverrides, Plyr]);
 
   if (!isVideoSupported) {
     return (